requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.0.1
emoji==2.2.0
et-xmlfile==1.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
numpy==1.24.1
openpyxl==3.0.10
pandas==1.5.3
python-dateutil==2.8.2
python-telegram-bot==20.0
pytz==2022.7.1
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
urllib3==1.26.14
